requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dateparser==0.7.6
discord.py==1.5.0
flake8==3.8.3
hiredis==1.1.0
idna==2.10
isort==5.5.0
mccabe==0.6.1
multidict==4.7.6
mypy-extensions==0.4.3
namegenerator==1.0.6
names==0.3.0
pathspec==0.8.0
psutil==5.7.2
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.9.27
sentry-sdk==0.17.3
six==1.15.0
statcord.py==3.0.4
subprocess32==3.5.4
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
yarl==1.6.0
